requirements.txt
customtkinter
python-mpv
yt-dlp
